Move signup bottom sheet to signup package
